==> scripts/07-replications.R <==
#### Preamble ####
# Purpose: Replicated graphs from... [...UPDATE THIS...]
# Author: Gace Nguyen 
# Date: 22 October 2024 
# Contact: hagiang.nguyen@mail.utoronto.ca C
# License: MIT
# Pre-requisites: [...UPDATE THIS...]
# Any other information needed? [...UPDATE THIS...]


#### Workspace setup ####
library(tidyverse)
# [...UPDATE THIS...]

#### Load data ####
# [...ADD CODE HERE...]
